added missing to run ufg option. added search for metadata file and write metadata file

+++ run.py
@@ -4,10 +4,6 @@ import subprocess
 import datetime
 from zipfile import ZipFile
 
-# run UDM
-subprocess.run(['python', '-m',  'openudm', '/data/inputs'])
-print('*** Ran UDM ***')
-
 from os import getenv, walk, mkdir, remove, listdir
 from os.path import join, isdir, isfile
 
@@ -24,7 +20,31 @@ def zip_file(path, file):
     os.remove(join(path, file))
     return
 
+def find_metadata_files():
+    """
+    Search all directories for any metadata files (metadat.csv)
+    """
 
+    suitable_extension_types = ['csv','']
+
+    files_ = []
+
+    for root, dirs, files in walk('/data/inputs'):
+        print(root, files)
+        for file in files:
+
+            # check if extension acceptable
+            extension = file.split('.')[-1]
+            print('checking extension of:', file)
+            if extension in suitable_extension_types:
+                # check if metadata text in file name
+                if 'metadata' in file:
+                    # file is good and what we are looking for
+                    files_.append(join(root, file))
+
+    print(files_)
+    return files_
+    
 def find_files():
     """
     Search all directories for any input files
@@ -47,15 +67,53 @@ def find_files():
     print(input_files)
     return input_files
 
-print('*** Output files ***')
-print(find_files())
-
-
-# move files to output dir
+# setting file paths
 result_data_dir = '/data/inputs'
+output_dir = '/data/outputs'
 output_data_dir = '/data/outputs/data'
 buildings_data_dir = '/data/outputs/buildings'
 
+# get environment values
+# fetch UFG value
+run_ufg = getenv('RUN_UFG')
+if run_ufg is None:
+    run_ufg = False
+elif run_ufg.lower() == 'true' or run_ufg == True:
+    run_ufg = True
+else:
+    run_ufg = False
+
+# check for any meta data files
+metadata_files = find_metadata_files() #search for any existing metadata files (expect to find at least one from the population data)
+print('Metadata files', metadata_files)
+
+# write a metadata file recording key parameters
+year = None
+ssp = None
+print('Saving metadata file')
+if len(metadata_files) == 1: # if one metadata file found
+    df = pd.read_csv(metadata_files[0]) # read in the file into a dataframe
+    print(df.head())
+    df.to_csv(join(data_path, output_dir, 'metadata.csv')) # write datadrame to csv
+    
+    # set some parameters which can be used later
+    year = df.loc[df['PARAMETER'] == 'YEAR', 'VALUE']
+    ssp = df.loc[df['PARAMETER'] == 'SSP', 'VALUE']
+    
+elif len(metadata_files) == 0:
+    print('No metadata files found')
+else:
+    print('Multiple metadata files found. This functionality has not been addded yet')
+
+
+# run UDM
+subprocess.run(['python', '-m',  'openudm', '/data/inputs'])
+print('*** Ran UDM ***')
+
+print('*** Output files ***')
+print(find_files())
+
+# move files to output dir
 # make output dir if not exists
 os.makedirs(output_data_dir, exist_ok=True)
 
@@ -66,34 +124,40 @@ for file_name in files_to_copy:
 
 print('*** Moved UDM output files ***')
 
-# run urban fabric generator tool
-# make output dir if not exists
-os.makedirs(buildings_data_dir, exist_ok=True)
-urban_fabric_raster = os.path.join(output_data_dir, 'out_uf.asc')
-subprocess.run(['generate_urban_fabric', '-i', '/data/outputs/data/out_cell_dph.asc', '-o', urban_fabric_raster])
+# if we want to rename any file output by UDM do this here.....
+#
+#
 
-print('*** Ran UFG ***')
+# run ufg tool is set by user
+if run_ufg:
+    # run urban fabric generator tool
+    # make output dir if not exists
+    os.makedirs(buildings_data_dir, exist_ok=True)
+    urban_fabric_raster = os.path.join(output_data_dir, 'out_uf.asc')
+    subprocess.run(['generate_urban_fabric', '-i', '/data/outputs/data/out_cell_dph.asc', '-o', urban_fabric_raster])
 
-# run raster to vector tool
-subprocess.run(['raster_to_vector', '-i', '/data/outputs/data/out_uf.asc', '-o',
-                os.path.join(buildings_data_dir, "urban_fabric.gpkg"), '-f' 'buildings,roads,greenspace'])
+    print('*** Ran UFG ***')
 
-# in an old version the data is stored in the wrong place. zip into a suitable output location
-zipObj = ZipFile('/data/outputs/data/urban_fabric.zip', 'w')
+    # run raster to vector tool
+    subprocess.run(['raster_to_vector', '-i', '/data/outputs/data/out_uf.asc', '-o',
+                    os.path.join(buildings_data_dir, "urban_fabric.gpkg"), '-f' 'buildings,roads,greenspace'])
 
-print('searching for output')
-for root, dirs, files in walk('/'):
-    #print(root, files)
-    for file in files:
-        file_extension = file.split('.')[-1]
-        if file_extension == 'gpkg':
-            if 'buildings.gpkg' or 'roads.gpkg' or 'greenspace.gpkg' or 'urban_fabric.gpkg' in file:
-                print(join(root, file))
-                zipObj.write(join(root, file))
-                os.remove(join(root, file))
-zipObj.close()
+    # in an old version the data is stored in the wrong place. zip into a suitable output location
+    zipObj = ZipFile('/data/outputs/data/urban_fabric.zip', 'w')
 
-# to save disk space, zip out_uf.asc and delete the raw file
-zip_file(output_data_dir, 'out_uf.asc')
+    print('searching for output')
+    for root, dirs, files in walk('/'):
+        #print(root, files)
+        for file in files:
+            file_extension = file.split('.')[-1]
+            if file_extension == 'gpkg':
+                if 'buildings.gpkg' or 'roads.gpkg' or 'greenspace.gpkg' or 'urban_fabric.gpkg' in file:
+                    print(join(root, file))
+                    zipObj.write(join(root, file))
+                    os.remove(join(root, file))
+    zipObj.close()
 
-print('*** Ran R2V ***')
+    # to save disk space, zip out_uf.asc and delete the raw file
+    zip_file(output_data_dir, 'out_uf.asc')
+
+    print('*** Ran R2V ***')
